+++ src/extension.ts
@@ -1,8 +1,14 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
-import formatterService from "./Services/FormatterService";
 import InformationMessage from "./Models/InformationMessage";
+import variableFormatter from "./Services/VariableFormatter";
+import stringFormatter from "./Services/StringFormatter";
+import escapeFormatter from "./Services/EscapeFormatter";
+import {
+  languageEscapeRules,
+  languageUnescapeRules,
+} from "./Models/EscapeRules";
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -15,80 +21,264 @@ export function activate(context: vscode.ExtensionContext) {
   // Now provide the implementation of the command with registerCommand
   // The commandId parameter must match the command field in package.json
 
-  // * Format to string, escape only `\` and `"`
-  const formatToStringDisposable = vscode.commands.registerCommand(
-    "smartFormatter.formatToString",
+  // #region String formatter
+
+  // * Format string to upper case
+  const toUpperCaseDisposable = vscode.commands.registerCommand(
+    "stringFormatter.toUpperCase",
     () => {
-      console.log(`Format to string started`);
+      console.log(`String formatter: To upper case started`);
 
       try {
-        formatterService.formatToString();
-      } catch (msg) {
-        if (msg instanceof InformationMessage) {
-          vscode.window.showInformationMessage(msg.message);
-        } else if (msg instanceof Error) {
-          vscode.window.showErrorMessage(msg.message);
-        }
+        stringFormatter.toUpperCase();
+      } catch (catchable) {
+        processCatchable(catchable);
       }
     }
   );
-  context.subscriptions.push(formatToStringDisposable);
+  context.subscriptions.push(toUpperCaseDisposable);
 
-  // * Format from string, remove escape for `\` and `"`
-  const formatFromStringDisposable = vscode.commands.registerCommand(
-    "smartFormatter.formatFromString",
+  // * Format string to lower case
+  const toLowerCaseDisposable = vscode.commands.registerCommand(
+    "stringFormatter.toLowerCase",
     () => {
-      console.log(`Format from string started`);
+      console.log(`String formatter: To lower case started`);
 
       try {
-        formatterService.formatFromString();
-      } catch (msg) {
-        if (msg instanceof InformationMessage) {
-          vscode.window.showInformationMessage(msg.message);
-        } else if (msg instanceof Error) {
-          vscode.window.showErrorMessage(msg.message);
-        }
+        stringFormatter.toLowerCase();
+      } catch (catchable) {
+        processCatchable(catchable);
       }
     }
   );
-  context.subscriptions.push(formatFromStringDisposable);
+  context.subscriptions.push(toLowerCaseDisposable);
 
-  // * Escape all character by programming language
+  // * Capitalize first character of each sentence
+  const toSentenceCaseDisposable = vscode.commands.registerCommand(
+    "stringFormatter.toSentenceCase",
+    () => {
+      console.log(`String formatter: To Sentence Case started`);
+
+      try {
+        stringFormatter.toSentenceCase();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(toSentenceCaseDisposable);
+
+  // * Capitalize all first character for each word
+  const toTitleCaseDisposable = vscode.commands.registerCommand(
+    "stringFormatter.toTitleCase",
+    () => {
+      console.log(`String formatter: To title case started`);
+
+      try {
+        stringFormatter.toTitleCase();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(toTitleCaseDisposable);
+
+  // * Format string into single line
+  const toSingleLineDisposable = vscode.commands.registerCommand(
+    "stringFormatter.toSingleLine",
+    () => {
+      console.log(`String formatter: To single line started`);
+
+      try {
+        stringFormatter.toSingleLine();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(toSingleLineDisposable);
+
+  // #endregion
+
+  // #region Variable formatter
+
+  // * Format string to variable, connect word with underscore
+  const toUnderscoreVariable = vscode.commands.registerCommand(
+    "variableFormatter.toUnderscoreVariable",
+    () => {
+      console.log(`Variable formatter: To underscore variable started`);
+
+      try {
+        variableFormatter.toUnderscoreVariable();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(toUnderscoreVariable);
+
+  // * Format string to variable, capitalize each word
+  const toPascalCaseDisposable = vscode.commands.registerCommand(
+    "variableFormatter.toPascalCase",
+    () => {
+      console.log(`Variable formatter: To pascal case started`);
+
+      try {
+        variableFormatter.toPascalCase();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(toPascalCaseDisposable);
+
+  // * Format string to variable, capitalize each word but not the first
+  const toCamelCaseDisposable = vscode.commands.registerCommand(
+    "variableFormatter.toCamelCase",
+    () => {
+      console.log(`Variable formatter: To camel case started`);
+
+      try {
+        variableFormatter.toCamelCase();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(toCamelCaseDisposable);
+
+  // * Format string to custom variable, get settings from configuration
+  const toCustomVariableDisposable = vscode.commands.registerCommand(
+    "variableFormatter.toCustomVariable",
+    () => {
+      console.log(`Variable formatter: To custom variable started`);
+
+      try {
+        variableFormatter.toCustomVariable();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(toCustomVariableDisposable);
+
+  // #endregion
+
+  // #region Escape formatter
+
+  // * Escape all characters by current file language id
   const escapeDisposable = vscode.commands.registerCommand(
-    "smartFormatter.escape",
+    "escapeFormatter.escape",
     () => {
-      console.log(`Escape started`);
+      console.log(`Escape formatter: Escape`);
 
       try {
-      } catch (msg) {
-        if (msg instanceof InformationMessage) {
-          vscode.window.showInformationMessage(msg.message);
-        } else if (msg instanceof Error) {
-          vscode.window.showErrorMessage(msg.message);
-        }
+        escapeFormatter.escapeWithCurrentLanguage(languageEscapeRules);
+      } catch (catchable) {
+        processCatchable(catchable);
       }
     }
   );
   context.subscriptions.push(escapeDisposable);
 
-  // * Unescape all character by programming language
+  // * Unescape all characters by current file language id
   const unescapeDisposable = vscode.commands.registerCommand(
-    "smartFormatter.unescape",
+    "escapeFormatter.unescape",
     () => {
-      console.log(`Unescape started`);
+      console.log(`Escape formatter: Unescape`);
 
       try {
-      } catch (msg) {
-        if (msg instanceof InformationMessage) {
-          vscode.window.showInformationMessage(msg.message);
-        } else if (msg instanceof Error) {
-          vscode.window.showErrorMessage(msg.message);
-        }
+        escapeFormatter.escapeWithCurrentLanguage(languageUnescapeRules);
+      } catch (catchable) {
+        processCatchable(catchable);
       }
     }
   );
   context.subscriptions.push(unescapeDisposable);
+
+  // * Escape all characters using the selected language id
+  const escapeWithLanguageDisposable = vscode.commands.registerCommand(
+    "escapeFormatter.escapeWithLanguage",
+    () => {
+      console.log(`Escape formatter: Escape with language`);
+
+      try {
+        escapeFormatter.escapeWithLanguage(languageEscapeRules);
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(escapeWithLanguageDisposable);
+
+  // * Unescape all characters using the selected language id
+  const unescapeWithLanguageDisposable = vscode.commands.registerCommand(
+    "escapeFormatter.unescapeWithLanguage",
+    () => {
+      console.log(`Escape formatter: Unescape with language`);
+
+      try {
+        escapeFormatter.escapeWithLanguage(languageUnescapeRules);
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(unescapeWithLanguageDisposable);
+
+  // * Format to string, escape only `\` and `"`
+  const toStringDisposable = vscode.commands.registerCommand(
+    "escapeFormatter.escapeStringValue",
+    () => {
+      console.log(`Escape formatter: Escape string value started`);
+
+      try {
+        variableFormatter.escapeStringValue();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(toStringDisposable);
+
+  // * Format from string, remove escape for `\` and `"`
+  const fromStringDisposable = vscode.commands.registerCommand(
+    "escapeFormatter.unescapeStringValue",
+    () => {
+      console.log(`Escape formatter: Unescape string value started`);
+
+      try {
+        variableFormatter.unescapeStringValue();
+      } catch (catchable) {
+        processCatchable(catchable);
+      }
+    }
+  );
+  context.subscriptions.push(fromStringDisposable);
+
+  // #endregion
+}
+
+function processCatchable(catchable: any) {
+  if (catchable instanceof InformationMessage) {
+    vscode.window.showInformationMessage(catchable.message);
+  } else if (catchable instanceof Error) {
+    vscode.window.showErrorMessage(catchable.message);
+  }
 }
 
 // this method is called when your extension is deactivated
 export function deactivate() {}
+// "menus": {
+//   "editor/context": [{
+//       "command": "toString",
+//       "group": "1_smartFormatter@0",
+//       "when": "config.editorContextMenu"
+//     },
+//     {
+//       "command": "fromString",
+//       "group": "1_smartFormatter@1",
+//       "when": "config.editorContextMenu"
+//     }
+//   ]
+// },